+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Yii 2 Install extension Change Log
 1.0.1 under development
 -----------------------
 
+- Enh #3: `InitController::$commands` now allows usage of callable (klimov-paul)
 - Bug #2: `InitController::actionRequirements()` prevents installation on warning in case requirements checking is performed by output analyzes (klimov-paul)